refactor: split main.go into focused files, delete demo code, fix storage race
Extract CLI commands (main, getOcapRecording, reduceMission, testQuery)
to cli.go and storage initialization (initStorage, getPostgresDB,
createStorageBackend) to storage.go. Delete dead demo data generation
code (~350 lines). Fix race condition where startGoroutines() created
the GORM storage backend using DB before :INIT:STORAGE: set it — now
initStorage() creates the DB connection first, then the backend.
Replace math.Max with Go 1.21+ builtin max().
