+++ cmd/ocap_recorder/main.go
@@ -8,34 +8,23 @@ package main
 import "C" // This is required to import the C code
 
 import (
-	"compress/gzip"
 	"context"
-	"encoding/json"
 	"fmt"
 	"log/slog"
-	"math"
-	"math/rand"
 	"os"
 	"path/filepath"
 	"runtime"
-	"strconv"
-	"strings"
 	"sync"
 	"time"
 
 	"github.com/OCAP2/extension/v5/internal/api"
 	"github.com/OCAP2/extension/v5/internal/cache"
 	"github.com/OCAP2/extension/v5/internal/config"
-	"github.com/OCAP2/extension/v5/internal/database"
 	"github.com/OCAP2/extension/v5/internal/dispatcher"
 	"github.com/OCAP2/extension/v5/internal/logging"
-	"github.com/OCAP2/extension/v5/internal/model"
 	intOtel "github.com/OCAP2/extension/v5/internal/otel"
 	"github.com/OCAP2/extension/v5/internal/parser"
 	"github.com/OCAP2/extension/v5/internal/storage"
-	gormstorage "github.com/OCAP2/extension/v5/internal/storage/gorm"
-	"github.com/OCAP2/extension/v5/internal/storage/memory"
-	sqlitestorage "github.com/OCAP2/extension/v5/internal/storage/sqlite"
 	"github.com/OCAP2/extension/v5/internal/util"
 	"github.com/OCAP2/extension/v5/internal/worker"
 	"github.com/OCAP2/extension/v5/pkg/a3interface"
@@ -88,9 +77,6 @@ var (
 
 	// OTelProvider handles OpenTelemetry
 	OTelProvider *intOtel.Provider
-
-	// testing
-	IsDemoData bool = false
 
 	// EntityCache is a map of all entities in the current mission, used to find associated entities by ocapID for entity state processing
 	EntityCache *cache.EntityCache = cache.NewEntityCache()
@@ -236,14 +222,12 @@ func init() {
 	Logger.Debug("Number of CPUs", "numCPUs", numCPUs)
 
 	// set GOMAXPROCS
-	runtime.GOMAXPROCS(int(math.Max(float64(numCPUs-2), 1)))
+	runtime.GOMAXPROCS(max(numCPUs-2, 1))
 
-	go func() {
-		startGoroutines()
+	// Initialize parser (no DB dependency)
+	parserService = parser.NewParser(Logger)
 
-		// Initialize API client and check frontend status
-		initAPIClient()
-	}()
+	go initAPIClient()
 }
 
 func initExtension() {
@@ -251,44 +235,6 @@ func initExtension() {
 	a3interface.WriteArmaCallback(ExtensionName, ":EXT:READY:")
 	// send extension version
 	a3interface.WriteArmaCallback(ExtensionName, ":VERSION:", BuildVersion)
-}
-
-func initStorage() error {
-	Logger.Debug("Received :INIT:STORAGE: call")
-	functionName := ":INIT:STORAGE:"
-
-	storageCfg := config.GetStorageConfig()
-	switch storageCfg.Type {
-	case "memory":
-		Logger.Info("Memory storage mode initialized")
-		a3interface.WriteArmaCallback(ExtensionName, ":STORAGE:OK:", "memory")
-
-	case "sqlite":
-		Logger.Info("SQLite storage mode initialized")
-		a3interface.WriteArmaCallback(ExtensionName, ":STORAGE:OK:", "sqlite")
-
-	case "postgres", "gorm", "database":
-		var err error
-		DB, err = getPostgresDB()
-		if err != nil {
-			SlogManager.WriteLog(functionName, fmt.Sprintf(`Error connecting to database: %v`, err), "ERROR")
-			a3interface.WriteArmaCallback(ExtensionName, ":STORAGE:ERROR:", err.Error())
-			return err
-		}
-		if DB == nil {
-			err = fmt.Errorf("database connection is nil")
-			SlogManager.WriteLog(functionName, err.Error(), "ERROR")
-			a3interface.WriteArmaCallback(ExtensionName, ":STORAGE:ERROR:", err.Error())
-			return err
-		}
-		a3interface.WriteArmaCallback(ExtensionName, ":STORAGE:OK:", DB.Dialector.Name())
-
-	default:
-		return fmt.Errorf("unknown storage type: %s", storageCfg.Type)
-	}
-
-	storageReadyOnce.Do(func() { close(storageReady) })
-	return nil
 }
 
 func setupA3Interface() (err error) {
@@ -330,88 +276,6 @@ func initAPIClient() {
 	} else {
 		Logger.Info("OCAP Frontend is online")
 	}
-}
-
-///////////////////////
-// DATABASE OPS //
-///////////////////////
-
-func getPostgresDB() (db *gorm.DB, err error) {
-	Logger.Debug("Connecting to Postgres DB")
-	db, err = database.GetPostgresDBStandalone()
-	if err != nil {
-		return nil, err
-	}
-
-	// Test connection
-	sqlDB, err := db.DB()
-	if err != nil {
-		return nil, fmt.Errorf("failed to access sql interface: %w", err)
-	}
-	if err = sqlDB.Ping(); err != nil {
-		return nil, fmt.Errorf("failed to validate connection: %w", err)
-	}
-
-	sqlDB.SetMaxOpenConns(10)
-	SlogManager.WriteLog(":INIT:DB:", "Connected to Postgres database", "INFO")
-	return db, nil
-}
-
-func startGoroutines() (err error) {
-	functionName := "startGoroutines"
-
-	// Initialize handler service
-	parserService = parser.NewParser(Logger)
-
-	// Initialize storage backend based on config
-	storageCfg := config.GetStorageConfig()
-	switch storageCfg.Type {
-	case "memory":
-		storageBackend = memory.New(storageCfg.Memory)
-		storageBackend.Init()
-		Logger.Info("Memory storage backend initialized")
-
-	case "sqlite":
-		sqliteDBFilePath := filepath.Join(AddonFolder, fmt.Sprintf("%s_%s.db", ExtensionName, SessionStartTime.Format("20060102_150405")))
-		backend, err := sqlitestorage.New(sqlitestorage.Config{
-			DumpInterval: storageCfg.SQLite.DumpInterval,
-			DumpPath:     sqliteDBFilePath,
-		}, EntityCache, MarkerCache, SlogManager)
-		if err != nil {
-			Logger.Error("Failed to create SQLite backend", "error", err)
-			return err
-		}
-		storageBackend = backend
-		storageBackend.Init()
-		Logger.Info("SQLite storage backend initialized")
-
-	default: // postgres, gorm, database
-		storageBackend = gormstorage.New(gormstorage.Dependencies{
-			DB:          DB,
-			EntityCache: EntityCache,
-			MarkerCache: MarkerCache,
-			LogManager:  SlogManager,
-		})
-		storageBackend.Init()
-		Logger.Info("GORM storage backend initialized")
-	}
-
-	// Initialize worker manager
-	workerManager = worker.NewManager(worker.Dependencies{
-		EntityCache:   EntityCache,
-		MarkerCache:   MarkerCache,
-		LogManager:    SlogManager,
-		ParserService: parserService,
-	}, storageBackend)
-
-	// Register worker handlers with the early dispatcher (created in setupA3Interface)
-	Logger.Debug("Registering worker handlers with dispatcher")
-	workerManager.RegisterHandlers(eventDispatcher)
-	Logger.Info("Worker handlers registered with dispatcher")
-
-	// log post goroutine creation
-	SlogManager.WriteLog(functionName, "Goroutines started successfully", "INFO")
-	return nil
 }
 
 // handleNewMission handles the :NEW:MISSION: command: parses mission data,
@@ -540,748 +404,4 @@ func registerLifecycleHandlers(d *dispatcher.Dispatcher) {
 		}
 		return nil, nil
 	})
-}
-
-// dispatchDemoEvent dispatches an event through the dispatcher for demo/test purposes
-func dispatchDemoEvent(command string, args []string) {
-	if eventDispatcher == nil {
-		return
-	}
-	eventDispatcher.Dispatch(dispatcher.Event{
-		Command:   command,
-		Args:      args,
-		Timestamp: time.Now(),
-	})
-}
-
-func populateDemoData() {
-	if DB == nil {
-		return
-	}
-
-	// declare test size counts
-	var (
-		numMissions        int = 2
-		missionDuration    int = 60 * 15
-		numUnitsPerMission int = 60
-		numSoldiers        int = int(math.Ceil(float64(numUnitsPerMission) * float64(0.8)))
-		numVehicles        int = numUnitsPerMission - numSoldiers
-
-		waitGroup = sync.WaitGroup{}
-
-		sides []string = []string{"WEST", "EAST", "GUER", "CIV"}
-
-		vehicleClassnames []string = []string{
-			"B_MRAP_01_F", "B_MRAP_01_gmg_F", "B_MRAP_01_hmg_F",
-			"B_G_Offroad_01_armed_F", "B_G_Offroad_01_AT_F", "B_G_Offroad_01_F",
-			"B_G_Offroad_01_repair_F", "B_APC_Wheeled_01_cannon_F",
-			"B_APC_Tracked_01_AA_F", "B_APC_Tracked_01_CRV_F",
-			"B_APC_Tracked_01_rcws_F",
-		}
-
-		ocapTypes []string = []string{
-			"ship", "parachute", "heli", "plane", "truck",
-			"car", "apc", "tank", "staticMortar", "unknown",
-		}
-
-		roles []string = []string{
-			"Rifleman", "Team Leader", "Auto Rifleman", "Assistant Auto Rifleman",
-			"Grenadier", "Machine Gunner", "Assistant Machine Gunner", "Medic",
-			"Engineer", "Explosive Specialist", "Rifleman (AT)", "Rifleman (AA)", "Officer",
-		}
-
-		roleDescriptions []string = []string{
-			"Rifleman@Alpha", "Team Leader@Alpha", "Auto Rifleman@Alpha",
-			"Assistant Auto Rifleman@Alpha", "Grenadier@Alpha", "Machine Gunner@Alpha",
-			"Assistant Machine Gunner@Alpha", "Medic@Alpha",
-			"Rifleman@Bravo", "Team Leader@Bravo", "Auto Rifleman@Bravo",
-			"Assistant Auto Rifleman@Bravo", "Grenadier@Bravo", "Machine Gunner@Bravo",
-			"Assistant Machine Gunner@Bravo", "Medic@Bravo",
-		}
-
-		worldNames []string = []string{
-			"Altis", "Stratis", "VR", "Bootcamp_ACR", "Malden",
-			"ProvingGrounds_PMC", "Shapur_BAF", "Sara", "Sara_dbe1", "SaraLite",
-			"Woodland_ACR", "Chernarus", "Desert_E", "Desert_Island", "Intro", "Desert2",
-		}
-
-		demoAddons [][]any = [][]any{
-			{"Community Base Addons v3.15.1", 0},
-			{"Arma 3 Contact", 0},
-			{"Arma 3 Creator DLC: Global Mobilization - Cold War Germany", 0},
-			{"Arma 3 Tanks", 0},
-			{"Arma 3 Tac-Ops", 0},
-			{"Arma 3 Laws of War", 0},
-			{"Arma 3 Apex", 0},
-			{"Arma 3 Marksmen", "332350"},
-			{"Arma 3 Helicopters", "304380"},
-			{"Arma 3 Karts", "288520"},
-			{"Arma 3 Zeus", "275700"},
-		}
-	)
-
-	// write worlds
-	missionsStart := time.Now()
-	for i := 0; i < numMissions; i++ {
-		data := make([]string, 2)
-
-		worldNameOriginal := worldNames[rand.Intn(len(worldNames))]
-		displayName := strings.Replace(worldNameOriginal, "_", " ", -1)
-		worldName := strings.ToLower(worldNameOriginal)
-
-		worldData := map[string]any{
-			"author":            "Demo Author",
-			"workshopID":        "123456789",
-			"displayName":       displayName,
-			"worldName":         worldName,
-			"worldNameOriginal": worldNameOriginal,
-			"worldSize":         10240,
-			"latitude":          rand.Float64() * 180,
-			"longitude":         rand.Float64() * 180,
-		}
-		worldDataJSON, err := json.Marshal(worldData)
-		if err != nil {
-			fmt.Println(err)
-		}
-		data[0] = string(worldDataJSON)
-
-		missionData := map[string]any{
-			"missionName":                  fmt.Sprintf("Demo Mission %d", i),
-			"briefingName":                 fmt.Sprintf("Demo Briefing %d", i),
-			"missionNameSource":            fmt.Sprintf("Demo Mission %d", i),
-			"onLoadName":                   "TestLoadName",
-			"author":                       "Demo Author",
-			"serverName":                   "Demo Server",
-			"serverProfile":                "Demo Profile",
-			"missionStart":                 nil,
-			"worldName":                    fmt.Sprintf("demo_world_%d", i),
-			"tag":                          "Demo Tag",
-			"captureDelay":                 1.0,
-			"addonVersion":                 "1.0",
-			"extensionVersion":             "1.0",
-			"extensionBuild":               "1.0",
-			"ocapRecorderExtensionVersion": "1.0",
-			"addons":                       demoAddons,
-			"playableSlots":                []float64{20, 20, 20, 5, 2},
-			"sideFriendly":                 []bool{false, false, true},
-		}
-		missionDataJSON, err := json.Marshal(missionData)
-		if err != nil {
-			fmt.Println(err)
-		}
-		data[1] = string(missionDataJSON)
-
-		dispatchDemoEvent(":NEW:MISSION:", data)
-
-		time.Sleep(500 * time.Millisecond)
-	}
-	missionsEnd := time.Now()
-	missionsElapsed := missionsEnd.Sub(missionsStart)
-	fmt.Printf("Sent %d missions in %s\n", numMissions, missionsElapsed)
-
-	// for each mission
-	missions := []model.Mission{}
-	DB.Model(&model.Mission{}).
-		Order("created_at DESC").
-		Limit(numMissions).
-		Find(&missions)
-
-	waitGroup = sync.WaitGroup{}
-	for _, mission := range missions {
-		if setter, ok := storageBackend.(interface{ SetMissionID(uint) }); ok {
-			setter.SetMissionID(mission.ID)
-		}
-
-		fmt.Printf("Populating mission with ID %d\n", mission.ID)
-
-		// write soldiers
-		idCounter := 1
-
-		for i := 0; i <= numSoldiers; i++ {
-			waitGroup.Add(1)
-			go func(thisId int) {
-				soldierID := strconv.FormatInt(int64(thisId), 10)
-				squadParams := []any{
-					[]string{"DD", "Diamond Dogs", "t@example.com", "https://example.com", "", "Diamond Dogs"},
-					[]string{"12491654", "Sleeping Tiger", "Sleeping Tiger", "st@example.com", "", "Thanks, Boss!"},
-					"1294510750",
-					"04328572",
-				}
-				squadParamsJSON, err := json.Marshal(squadParams)
-				if err != nil {
-					fmt.Println(err)
-				}
-
-				soldier := []string{
-					soldierID,
-					soldierID,
-					fmt.Sprintf("Demo Unit %d", i),
-					fmt.Sprintf("Demo Group %d", i),
-					sides[rand.Intn(len(sides))],
-					strconv.FormatBool(rand.Intn(2) == 1),
-					roleDescriptions[rand.Intn(len(roleDescriptions))],
-					"B_Soldier_F",
-					"Rifleman",
-					strconv.FormatInt(int64(rand.Intn(1000000000)), 10),
-					string(squadParamsJSON),
-				}
-
-				dispatchDemoEvent(":NEW:SOLDIER:", soldier)
-
-				for {
-					time.Sleep(100 * time.Millisecond)
-					if _, ok := EntityCache.GetSoldier(uint16(thisId)); ok {
-						break
-					}
-				}
-
-				var randomPos [3]float64 = [3]float64{rand.Float64() * 30720, rand.Float64() * 30720, rand.Float64() * 30720}
-				var randomDir float64 = rand.Float64() * 360
-				var dirMoveOffset float64 = rand.Float64() * 360
-				var randomLifestate int = rand.Intn(3)
-				var currentRole string = roles[rand.Intn(len(roles))]
-				for j := 0; j <= missionDuration; j++ {
-					time.Sleep(100 * time.Millisecond)
-
-					stateFrame := j
-					var xyTransform [2]float64 = [2]float64{0, 0}
-					if randomDir < 180 {
-						xyTransform[0] = math.Sin((randomDir + dirMoveOffset) * (math.Pi / 180))
-						xyTransform[1] = math.Cos((randomDir + dirMoveOffset) * (math.Pi / 180))
-					} else {
-						xyTransform[0] = math.Sin((randomDir - dirMoveOffset) * (math.Pi / 180))
-						xyTransform[1] = math.Cos((randomDir - dirMoveOffset) * (math.Pi / 180))
-					}
-					randomPos[0] += xyTransform[0] + (rand.Float64() * 8) - 4
-					randomPos[1] += xyTransform[1] + (rand.Float64() * 8) - 4
-					randomPos[2] += (rand.Float64() * 8) - 4
-					randomDir += (rand.Float64() * 8) - 4
-					if rand.Intn(10) == 0 {
-						randomLifestate = rand.Intn(3)
-					}
-					if rand.Intn(20) == 0 {
-						currentRole = roles[rand.Intn(len(roles))]
-					}
-
-					soldierState := []string{
-						soldierID,
-						fmt.Sprintf("[%f,%f,%f]", randomPos[0], randomPos[1], randomPos[2]),
-						strconv.FormatFloat(randomDir, 'f', 0, 64),
-						strconv.FormatInt(int64(randomLifestate), 10),
-						strconv.FormatBool(rand.Intn(2) == 1),
-						fmt.Sprintf("Demo Unit %d", j),
-						strconv.FormatBool(rand.Intn(2) == 1),
-						currentRole,
-						strconv.FormatInt(int64(stateFrame), 10),
-						strconv.FormatInt(int64(rand.Intn(2)), 10),
-						strconv.FormatInt(int64(rand.Intn(2)), 10),
-						fmt.Sprintf("%d,%d,%d,%d,%d,%d", rand.Intn(20), rand.Intn(20), rand.Intn(20), rand.Intn(20), rand.Intn(20), rand.Intn(20)),
-						"Passenger",
-						"-1",
-						[]string{"Up", "Middle", "Down"}[rand.Intn(3)],
-					}
-
-					dispatchDemoEvent(":NEW:SOLDIER:STATE:", soldierState)
-				}
-				waitGroup.Done()
-			}(idCounter)
-			idCounter++
-		}
-
-		Logger.Debug("Soldiers cached", "numSoldiersCached", EntityCache.SoldierCount())
-
-		// write vehicles
-		for i := 0; i <= numVehicles; i++ {
-			waitGroup.Add(1)
-			go func(thisId int) {
-				vehicleID := strconv.FormatInt(int64(thisId), 10)
-				vehicle := []string{
-					vehicleID,
-					vehicleID,
-					ocapTypes[rand.Intn(len(ocapTypes))],
-					fmt.Sprintf("Demo Vehicle %d", i),
-					vehicleClassnames[rand.Intn(len(vehicleClassnames))],
-					fmt.Sprintf(`"[[""%s"", %d], [""%s"", %d], [""%s"", %d]]"`, "wasp", 1, "AddTread", 1, "AddTread_Short", 1),
-				}
-
-				dispatchDemoEvent(":NEW:VEHICLE:", vehicle)
-
-				for {
-					time.Sleep(1000 * time.Millisecond)
-					if _, ok := EntityCache.GetVehicle(uint16(thisId)); ok {
-						break
-					}
-				}
-
-				for j := 0; j <= missionDuration; j++ {
-					time.Sleep(100 * time.Millisecond)
-					vehicleState := []string{
-						vehicleID,
-						fmt.Sprintf("[%f,%f,%f]", rand.Float64()*30720+1, rand.Float64()*30720+1, rand.Float64()*30720+1),
-						fmt.Sprintf("%d", rand.Intn(360)),
-						strconv.FormatBool(rand.Intn(2) == 1),
-						fmt.Sprintf("[%d,%d,%d]", rand.Intn(numSoldiers)+1, rand.Intn(numSoldiers)+1, rand.Intn(numSoldiers)+1),
-						strconv.FormatInt(int64(rand.Intn(missionDuration)), 10),
-						fmt.Sprintf("%f", rand.Float64()),
-						fmt.Sprintf("%f", rand.Float64()),
-						strconv.FormatBool(rand.Intn(10) == 0),
-						strconv.FormatBool(rand.Intn(10) == 0),
-						sides[rand.Intn(len(sides))],
-						fmt.Sprintf("[%f,%f,%f]", rand.Float64(), rand.Float64(), rand.Float64()),
-						fmt.Sprintf("[%f,%f,%f]", rand.Float64(), rand.Float64(), rand.Float64()),
-						fmt.Sprintf("%f", rand.Float64()),
-						fmt.Sprintf("%f", rand.Float64()),
-					}
-
-					dispatchDemoEvent(":NEW:VEHICLE:STATE:", vehicleState)
-				}
-				waitGroup.Done()
-			}(idCounter)
-			idCounter++
-		}
-
-		waitGroup.Wait()
-
-		fmt.Println("Finished creating units and states.")
-
-		// demo markers
-		markerTypes := []string{"mil_dot", "mil_triangle", "mil_box", "hd_flag"}
-		markerColors := []string{"ColorRed", "ColorBlue", "ColorGreen", "ColorYellow"}
-		markerShapes := []string{"ICON", "RECTANGLE", "ELLIPSE"}
-
-		for i := 0; i < 10; i++ {
-			marker := []string{
-				fmt.Sprintf("DemoMarker_%d", i),
-				fmt.Sprintf("%d", rand.Intn(360)),
-				markerTypes[rand.Intn(len(markerTypes))],
-				fmt.Sprintf("Demo Marker %d", i),
-				strconv.Itoa(rand.Intn(missionDuration)),
-				"-1",
-				strconv.Itoa(rand.Intn(numSoldiers) + 1),
-				markerColors[rand.Intn(len(markerColors))],
-				"[1,1]",
-				sides[rand.Intn(len(sides))],
-				fmt.Sprintf("[%f,%f,%f]", rand.Float64()*30720+1, rand.Float64()*30720+1, 0.0),
-				markerShapes[rand.Intn(len(markerShapes))],
-				"1.0",
-				"Solid",
-			}
-			dispatchDemoEvent(":MARKER:CREATE:", marker)
-
-			for j := 0; j < 3; j++ {
-				time.Sleep(100 * time.Millisecond)
-				markerMove := []string{
-					fmt.Sprintf("DemoMarker_%d", i),
-					strconv.Itoa(rand.Intn(missionDuration)),
-					fmt.Sprintf("[%f,%f,%f]", rand.Float64()*30720+1, rand.Float64()*30720+1, 0.0),
-					fmt.Sprintf("%d", rand.Intn(360)),
-					"1.0",
-				}
-				dispatchDemoEvent(":MARKER:MOVE:", markerMove)
-			}
-		}
-
-		for i := 0; i < 3; i++ {
-			markerDelete := []string{
-				fmt.Sprintf("DemoMarker_%d", i),
-				strconv.Itoa(missionDuration - 10),
-			}
-			dispatchDemoEvent(":MARKER:DELETE:", markerDelete)
-		}
-	}
-
-	// Give dispatcher time to process buffered events
-	time.Sleep(2 * time.Second)
-}
-
-func getOcapRecording(missionIDs []string) (err error) {
-	fmt.Println("Getting JSON for mission IDs: ", missionIDs)
-
-	for _, missionID := range missionIDs {
-		missionIDInt, err := strconv.Atoi(missionID)
-		if err != nil {
-			return err
-		}
-
-		txStart := time.Now()
-		var mission model.Mission
-		ocapMission := make(map[string]any)
-		err = DB.Model(&model.Mission{}).Where("id = ?", missionIDInt).First(&mission).Error
-		if err != nil {
-			return err
-		}
-
-		ocapMission["addonVersion"] = mission.AddonVersion
-		ocapMission["extensionVersion"] = mission.ExtensionVersion
-		ocapMission["extensionBuild"] = mission.ExtensionBuild
-		ocapMission["ocapRecorderExtensionVersion"] = mission.OcapRecorderExtensionVersion
-
-		ocapMission["missionAuthor"] = mission.Author
-		ocapMission["missionName"] = mission.OnLoadName
-		if mission.OnLoadName == "" {
-			ocapMission["missionName"] = mission.MissionName
-		}
-
-		world := model.World{}
-		err = DB.Model(&model.World{}).Where("id = ?", mission.WorldID).First(&world).Error
-		if err != nil {
-			return fmt.Errorf("error getting world: %w", err)
-		}
-		ocapMission["worldName"] = world.WorldName
-
-		totalSoldiers := int64(0)
-		err = DB.Model(&model.Soldier{}).Where("mission_id = ?", missionIDInt).Count(&totalSoldiers).Error
-		if err != nil {
-			return fmt.Errorf("error getting soldier count: %w", err)
-		}
-
-		totalVehicles := int64(0)
-		err = DB.Model(&model.Vehicle{}).Where("mission_id = ?", missionIDInt).Count(&totalVehicles).Error
-		if err != nil {
-			return fmt.Errorf("error getting vehicle count: %w", err)
-		}
-
-		ocapMission["Rone"] = map[string]any{}
-		ocapMission["events"] = []any{}
-
-		soldiers := []model.Soldier{}
-		soldierTxStart := time.Now()
-		err = DB.Model(&model.Soldier{}).Where("mission_id = ?", missionIDInt).Find(&soldiers).Error
-		if err != nil {
-			return fmt.Errorf("error getting soldiers: %w", err)
-		}
-		fmt.Println("Got soldiers in ", time.Since(soldierTxStart))
-
-		entities := []map[string]any{}
-		for _, soldier := range soldiers {
-			entity := map[string]any{}
-			entity["id"] = soldier.ObjectID
-			entity["name"] = soldier.UnitName
-			entity["group"] = soldier.GroupID
-			entity["side"] = soldier.Side
-			entity["isPlayer"] = 0
-			if soldier.IsPlayer {
-				entity["isPlayer"] = 1
-			}
-			entity["type"] = "unit"
-			entity["startFrameNum"] = soldier.JoinFrame
-
-			soldierStates := []model.SoldierState{}
-			err = DB.Model(&model.SoldierState{}).
-				Where("mission_id = ? AND soldier_ocap_id = ?", missionIDInt, soldier.ObjectID).
-				Order("capture_frame ASC").
-				Find(&soldierStates).Error
-			if err != nil {
-				return fmt.Errorf("error getting soldier states: %w", err)
-			}
-
-			positions := []any{}
-			for _, state := range soldierStates {
-				coord, _ := state.Position.Coordinates()
-				position := []any{
-					[]float64{coord.XY.X, coord.XY.Y},
-					state.Bearing,
-					state.Lifestate,
-					state.InVehicleObjectID,
-					state.UnitName,
-					state.IsPlayer,
-					state.CurrentRole,
-				}
-				positions = append(positions, position)
-			}
-			entity["positions"] = positions
-
-			projectileEvents := []model.ProjectileEvent{}
-			err = DB.Model(&model.ProjectileEvent{}).
-				Where("mission_id = ? AND firer_object_id = ?", missionIDInt, soldier.ObjectID).
-				Order("capture_frame ASC").
-				Find(&projectileEvents).Error
-			if err != nil {
-				return fmt.Errorf("error getting projectile events: %w", err)
-			}
-
-			framesFired := []any{}
-			for _, event := range projectileEvents {
-				var startX, startY, endX, endY float64
-				if !event.Positions.IsEmpty() {
-					if ls, ok := event.Positions.AsLineString(); ok {
-						seq := ls.Coordinates()
-						if seq.Length() > 0 {
-							start := seq.Get(0)
-							startX, startY = start.X, start.Y
-							end := seq.Get(seq.Length() - 1)
-							endX, endY = end.X, end.Y
-						}
-					}
-				}
-				frameFired := []any{
-					event.CaptureFrame,
-					[]float64{endX, endY},
-					[]float64{startX, startY},
-					event.WeaponDisplay,
-					event.MagazineDisplay,
-					event.Mode,
-				}
-				framesFired = append(framesFired, frameFired)
-			}
-			entity["framesFired"] = framesFired
-
-			entities = append(entities, entity)
-		}
-
-		vehicles := []model.Vehicle{}
-		err = DB.Model(&model.Vehicle{}).Where("mission_id = ?", missionIDInt).Find(&vehicles).Error
-		if err != nil {
-			return fmt.Errorf("error getting vehicles: %w", err)
-		}
-		for _, vehicle := range vehicles {
-			entity := map[string]any{}
-			entity["id"] = vehicle.ObjectID
-			entity["name"] = vehicle.DisplayName
-			entity["class"] = vehicle.ClassName
-			entity["side"] = "UNKNOWN"
-			entity["type"] = vehicle.OcapType
-			entity["startFrameNum"] = vehicle.JoinFrame
-
-			vehicleStates := []model.VehicleState{}
-			err = DB.Model(&model.VehicleState{}).
-				Where("mission_id = ? AND vehicle_ocap_id = ?", missionIDInt, vehicle.ObjectID).
-				Order("capture_frame ASC").
-				Find(&vehicleStates).Error
-			if err != nil {
-				return fmt.Errorf("error getting vehicle states: %w", err)
-			}
-			positions := []any{}
-			for _, state := range vehicleStates {
-				coord, _ := state.Position.Coordinates()
-				var crew any
-				if err := json.Unmarshal([]byte(state.Crew), &crew); err != nil {
-					crew = []any{}
-				}
-				position := []any{
-					[]float64{coord.XY.X, coord.XY.Y},
-					state.Bearing,
-					state.IsAlive,
-					crew,
-				}
-				positions = append(positions, position)
-			}
-			entity["positions"] = positions
-			entity["framesFired"] = []any{}
-
-			entities = append(entities, entity)
-		}
-
-		ocapMission["entities"] = entities
-
-		// Compute endFrame from the maximum capture_frame across all states
-		var endFrame uint
-		DB.Model(&model.SoldierState{}).Where("mission_id = ?", missionIDInt).Select("COALESCE(MAX(capture_frame), 0)").Scan(&endFrame)
-		ocapMission["endFrame"] = endFrame
-
-		fmt.Println("Got mission data in ", time.Since(txStart))
-
-		ocapMissionJSON, err := json.Marshal(ocapMission)
-		if err != nil {
-			return fmt.Errorf("error marshalling mission data: %w", err)
-		}
-
-		fileName := fmt.Sprintf("%s_%s.json.gz", mission.MissionName, mission.StartTime.Format("20060102_150405"))
-		fileName = strings.ReplaceAll(fileName, " ", "_")
-		fileName = strings.ReplaceAll(fileName, ":", "_")
-		f, err := os.Create(fileName)
-		if err != nil {
-			return fmt.Errorf("error creating file: %w", err)
-		}
-		defer f.Close()
-
-		gzWriter := gzip.NewWriter(f)
-		defer gzWriter.Close()
-		_, err = gzWriter.Write(ocapMissionJSON)
-		if err != nil {
-			return fmt.Errorf("error writing to gzip: %w", err)
-		}
-
-		fmt.Println("Wrote mission data to ", fileName)
-	}
-
-	return nil
-}
-
-func reduceMission(missionIDs []string) (err error) {
-	for _, missionID := range missionIDs {
-		missionIDInt, err := strconv.Atoi(missionID)
-		if err != nil {
-			return err
-		}
-
-		txStart := time.Now()
-		var mission model.Mission
-		err = DB.Model(&model.Mission{}).Where("id = ?", missionIDInt).First(&mission).Error
-		if err != nil {
-			return fmt.Errorf("error getting mission: %w", err)
-		}
-
-		soldierStatesToDelete := []model.SoldierState{}
-		err = DB.Model(&model.SoldierState{}).Where(
-			"mission_id = ? AND capture_frame % 5 != 0",
-			mission.ID,
-		).Order("capture_frame ASC").Find(&soldierStatesToDelete).Error
-		if err != nil {
-			return fmt.Errorf("error getting soldier states to delete: %w", err)
-		}
-
-		if len(soldierStatesToDelete) == 0 {
-			fmt.Println("No soldier states to delete for missionId ", missionID, ", checked in ", time.Since(txStart))
-			continue
-		}
-
-		err = DB.Delete(&soldierStatesToDelete).Error
-		if err != nil {
-			return fmt.Errorf("error deleting soldier states: %w", err)
-		}
-
-		fmt.Println("Deleted ", len(soldierStatesToDelete), " soldier states from missionId ", missionID, " in ", time.Since(txStart))
-	}
-
-	fmt.Println("")
-	fmt.Println("----------------------------------------")
-	fmt.Println("")
-	fmt.Println("Finished reducing soldier states, running VACUUM to recover space...")
-	txStart := time.Now()
-	tables := []string{}
-	err = DB.Raw(
-		`SELECT table_name FROM information_schema.tables WHERE table_schema = 'public' AND table_type = 'BASE TABLE'`,
-	).Scan(&tables).Error
-	if err != nil {
-		return fmt.Errorf("error getting tables to vacuum: %w", err)
-	}
-
-	for _, table := range tables {
-		err = DB.Raw(fmt.Sprintf("VACUUM (FULL) %s", table)).Error
-		if err != nil {
-			return fmt.Errorf("error running VACUUM on table %s: %w", table, err)
-		}
-	}
-
-	fmt.Println("Finished VACUUM in ", time.Since(txStart))
-	fmt.Println("Finished reducing, press enter to exit.")
-
-	return nil
-}
-
-func testQuery() (err error) {
-	query := `
-select
-    s.ocap_id,
-    ss.capture_frame,
-    json_agg(ss.*) as states,
-    json_agg(hit.*) as hits,
-    json_agg(kill.*) as kills,
-    json_agg(fire.*) as fired,
-    json_agg(re.*) as radio,
-    json_agg(ce.*) as chat
-from soldiers s
-  left join (
-    select *
-    from soldier_states
-    order by capture_frame asc
-  ) ss on ss.soldier_id = s.id
-  left join kill_events kill on (
-    kill.victim_id_soldier = s.id
-    or kill.killer_id_soldier = s.id
-  )
-  and ss.capture_frame = kill.capture_frame
-  left join hit_events hit on (
-    hit.victim_id_soldier = s.id
-    or hit.shooter_id_soldier = s.id
-  )
-  and ss.capture_frame = hit.capture_frame
-  left join fired_events fire on fire.soldier_id = s.id
-  and ss.capture_frame = fire.capture_frame
-  left join radio_events re on re.soldier_id = s.id
-  and ss.capture_frame = re.capture_frame
-  left join chat_events ce on ce.soldier_id = s.id
-  and ss.capture_frame = ce.capture_frame
-where s.mission_id = ? and ss.capture_frame between ? and ?
-group by s.ocap_id,
-  ss.capture_frame
-order by s.ocap_id,
-  ss.capture_frame;
-`
-
-	frameData := []model.FrameData{}
-	err = DB.Raw(query, 4, 0, 100).Scan(&frameData).Error
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-
-	jsonBytes, err := json.Marshal(frameData)
-	if err != nil {
-		return err
-	}
-	err = os.WriteFile("test.json", jsonBytes, 0644)
-	if err != nil {
-		return err
-	}
-
-	fmt.Println("Done!")
-	return nil
-}
-
-func main() {
-	var err error
-	Logger.Info("Starting up...")
-
-	Logger.Info("Initializing storage...")
-	err = initStorage()
-	if err != nil {
-		panic(err)
-	}
-	Logger.Info("Storage initialization complete.")
-	initExtension()
-
-	args := os.Args[1:]
-	if len(args) > 0 {
-		if strings.ToLower(args[0]) == "demo" {
-			Logger.Info("Populating demo data...")
-			IsDemoData = true
-			demoStart := time.Now()
-			populateDemoData()
-			Logger.Info("Demo data populated.", "duration", time.Since(demoStart))
-			fmt.Println("Press enter to exit.")
-		}
-		if strings.ToLower(args[0]) == "getjson" {
-			missionIds := args[1:]
-			if len(missionIds) > 0 {
-				err = getOcapRecording(missionIds)
-				if err != nil {
-					panic(err)
-				}
-			} else {
-				fmt.Println("No mission IDs provided.")
-			}
-		}
-		if strings.ToLower(args[0]) == "reducemission" {
-			missionIds := args[1:]
-			if len(missionIds) > 0 {
-				err = reduceMission(missionIds)
-				if err != nil {
-					panic(err)
-				}
-			} else {
-				fmt.Println("No mission IDs provided.")
-			}
-		}
-		if strings.ToLower(args[0]) == "testquery" {
-			err = testQuery()
-			if err != nil {
-				panic(err)
-			}
-		}
-	} else {
-		fmt.Println("No arguments provided.")
-	}
-	fmt.Scanln()
 }
